requirements.txt
certifi==2020.6.20
cryptography==3.1
hkp4py==0.2.3.1
pem==20.1.0
pyOpenSSL==17.5.0
urllib3==1.25.10